refactor: moved `CartItemDto` to separate class and used `BookRepository` method directly in `ShoppingCartServiceImpl` instead of DTO-to-entity conversion
