+++ src/main/java/mate/academy/bookstore/repository/shoppingcart/ShoppingCartRepository.java
@@ -6,8 +6,6 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
-    /*@Query("SELECT sc FROM ShoppingCart sc "
-            + "LEFT JOIN FETCH sc.cartItems WHERE sc.user.id = :userId AND sc.isDeleted = false")*/
     @EntityGraph(attributePaths = {"cartItems", "cartItems.book"})
     Optional<ShoppingCart> findByUserId(Long userId);
 }
